L3KanjiをL3KanjiDictとL3KanjiSelectの２つに分割。 その前提で、オブジェクトの生存期間を変更。 変換開始処理、変換処理を実装。 size_tのwarnが出ないように修正。

+++ hitomoji-engine/ChmConfig.cpp
@@ -207,8 +207,8 @@ BOOL ChmConfig::LoadFile(const std::wstring& fileName)
 
 	BOOL bRet = _LoadStreamInternal(ifs, path, /*isMain*/ TRUE, L"");
 
-	OutputDebugStringWithInt(L"   > LoadFile end. m_errors.size(): %d\n", m_errors.size());
-	OutputDebugStringWithInt(L"   >               m_infos.size(): %d\n", m_infos.size());
+	OutputDebugStringWithInt(L"   > LoadFile end. m_errors.size(): %d\n", (ULONG)m_errors.size());
+	OutputDebugStringWithInt(L"   >               m_infos.size(): %d\n", (ULONG)m_infos.size());
 
 	return bRet;
 }
